+++ app/src/main/java/com/survivalcoding/noteapp/NoteApplication.kt
@@ -2,7 +2,7 @@ package com.survivalcoding.noteapp
 
 import android.app.Application
 import androidx.room.Room
-import com.survivalcoding.noteapp.data.data_source.NoteDatabase
+import com.survivalcoding.noteapp.data.datasource.NoteDatabase
 import com.survivalcoding.noteapp.data.repository.NoteRoomRepository
 
 class NoteApplication : Application() {
