+++ app/src/main/java/com/survivalcoding/noteapp/domain/model/Note.kt
@@ -1,10 +1,11 @@
 package com.survivalcoding.noteapp.domain.model
 
 import androidx.room.Entity
+import androidx.room.PrimaryKey
 
 @Entity
 data class Note(
-    val id: Int? = null,
+    @PrimaryKey(autoGenerate = true) val id: Int? = null,
     val title: String,
     val content: String,
     val timestamp: Long,
