+++ app/src/main/java/com/survivalcoding/noteapp/domain/repository/NoteRepository.kt
@@ -3,7 +3,7 @@ package com.survivalcoding.noteapp.domain.repository
 import com.survivalcoding.noteapp.domain.model.Note
 
 interface NoteRepository: SelectAllRepository<Note> {
-    suspend fun getById(id: Int): Note?
+    suspend fun selectById(id: Int): Note?
 
     suspend fun upsert(note: Note)
