+++ app/src/main/java/com/survivalcoding/noteapp/presentation/note/NoteViewModel.kt
@@ -1,4 +1,4 @@
-package com.survivalcoding.noteapp.presentation.add_edit_note
+package com.survivalcoding.noteapp.presentation.note
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
@@ -8,14 +8,14 @@ import com.survivalcoding.noteapp.domain.model.Note
 import com.survivalcoding.noteapp.domain.repository.NoteRepository
 import kotlinx.coroutines.launch
 
-class AddEditNoteViewModel(private val noteRepository: NoteRepository): ViewModel() {
+class NoteViewModel(private val noteRepository: NoteRepository): ViewModel() {
 
-    private val _note = MutableLiveData<Note>()
-    val note: LiveData<Note> = _note
+    private val _note = MutableLiveData<Note?>()
+    val note: LiveData<Note?> = _note
 
     fun getNoteById(id: Int) {
         viewModelScope.launch {
-            _note.value = noteRepository.getById(id)
+            _note.value = noteRepository.selectById(id)
         }
     }
